Update wallet.mdx
The unstake is not working so well just by adding 10. I think we should try boosting it to 30. This is mainly because the reflection of the ETL height and the real chain height and pending transactions being slow. If you use +10 you will get a lot of failed unstakes.. > 50% of the time.